[msbuild] Merge the Codesign[TaskBase] classes. (#19498)
