--- msbuild/Xamarin.MacDev.Tasks/Tasks/Codesign.cs
@@ -1,28 +1,0 @@
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.Build.Framework;
-using Xamarin.Messaging.Build.Client;
-
-namespace Xamarin.MacDev.Tasks {
-	public class Codesign : CodesignTaskBase, ITaskCallback, ICancelableTask {
-		public override bool Execute ()
-		{
-			if (ShouldExecuteRemotely ())
-				return new TaskRunner (SessionId, BuildEngine4).RunAsync (this).Result;
-
-			return base.Execute ();
-		}
-
-		public IEnumerable<ITaskItem> GetAdditionalItemsToBeCopied () => Enumerable.Empty<ITaskItem> ();
-
-		public bool ShouldCopyToBuildServer (ITaskItem item) => false;
-
-		public bool ShouldCreateOutputFile (ITaskItem item) => true;
-
-		public void Cancel ()
-		{
-			if (ShouldExecuteRemotely ())
-				BuildConnection.CancelAsync (BuildEngine4).Wait ();
-		}
-	}
-}
